// generated/backend/CreatedMessageWithContent.ts
/**
 * Do not edit this file it is auto-generated by italia-utils / gen-api-models.
 * See https://github.com/teamdigitale/italia-utils
 */
/* tslint:disable */

import { FiscalCode } from "./FiscalCode";
import { TimeToLiveSeconds } from "./TimeToLiveSeconds";
import { Timestamp } from "./Timestamp";
import { MessageContent } from "./MessageContent";
import * as t from "io-ts";

// required attributes
const CreatedMessageWithContentR = t.interface({
  id: t.string,

  fiscal_code: FiscalCode,

  created_at: Timestamp,

  content: MessageContent,

  sender_service_id: t.string
});

// optional attributes
const CreatedMessageWithContentO = t.partial({
  time_to_live: TimeToLiveSeconds
});

export const CreatedMessageWithContent = t.exact(
  t.intersection(
    [CreatedMessageWithContentR, CreatedMessageWithContentO],
    "CreatedMessageWithContent"
  )
);

export type CreatedMessageWithContent = t.TypeOf<
  typeof CreatedMessageWithContent
>;